remove duplicated code with new service
