+++ frontend/src/app/example-table/example-table.component.ts
@@ -15,6 +15,7 @@ import { MatTable } from '@angular/material/table';
 import { StepDefinition } from '../model/StepDefinition';
 import { ThemePalette } from '@angular/material/core';
 import { ThemingService } from '../Services/theming.service';
+import { HighlightInputService } from '../Services/highlight-input.service';
 
 
 @Component({
@@ -158,8 +159,8 @@ export class ExampleTableComponent implements OnInit {
        private toastr: ToastrService,
        public exampleService: ExampleService,
        public apiService: ApiService,
-       public themeService: ThemingService
-
+       public themeService: ThemingService,
+       public highlightInputService: HighlightInputService
        ) {}
 
      /**
@@ -427,9 +428,7 @@ export class ExampleTableComponent implements OnInit {
    * @param initialCall if call is from ngDoCheck
    */
   private highlightRegex(el, columnIndex, rowIndex, initialCall) {
-    const regex = /(\{Regex:)(.*?)(\})(?=\s|$)/g;// Regex pattern to recognize and highlight regex expressions -> start with {Regex: and end with }
     const inputValue: string = el.textContent;
-    const offset = this.getCaretCharacterOffsetWithin(el)
 
     if(!initialCall){
       this.selectedScenario.stepDefinitions.example[rowIndex + 1].values[columnIndex-1] = inputValue;
@@ -438,123 +437,14 @@ export class ExampleTableComponent implements OnInit {
     if(!initialCall){
       this.initialRegex = false;
     }
-      
-    let highlightedText: string;
-    var regexDetected = false;
-    if(this.isDark){
-      highlightedText = inputValue.replace(regex, (match, match1, match2, match3) => {
-        regexDetected = true;
-        return `<span uk-tooltip="title:Regular Expression detected!;pos:right">`+
-        `<span style="color: var(--light-grey); font-weight: bold">${match1}</span>`+
-        `<span style="color: var(--light-blue); font-weight: bold">${match2}</span>`+
-        `<span style="color: var(--light-grey); font-weight: bold">${match3}</span></span>`;
-      });
-    } else{
-      highlightedText = inputValue.replace(regex, (match, match1, match2, match3) => {
-        regexDetected = true;
-        return `<span uk-tooltip="title:Regular Expression detected!;pos:right">`+
-        `<span style="color: var(--brown-grey); font-weight: bold">${match1}</span>`+
-        `<span style="color: var(--ocean-blue); font-weight: bold">${match2}</span>`+
-        `<span style="color: var(--brown-grey); font-weight: bold">${match3}</span></span>`;
-      });
-    }
 
-    el.innerHTML = highlightedText
+    var regexDetected = false;
+    
+    regexDetected = this.highlightInputService.highlightRegex(el,initialCall, this.isDark, this.regexInStory)
 
     if(initialCall && regexDetected) {
       this.regexInStory = true
     }
-    if(regexDetected && !this.regexInStory){
-      this.regexInStory = true
-      this.toastr.info('View our Documentation for more Info','Regular Expression detected!');
-    }
-
-    // Set cursor to correct position
-    if(!initialCall){
-      if (regexDetected){ //maybe not needed
-        const selection = window.getSelection();
-        selection.removeAllRanges()
-
-        // Call the function to find the correct node and offset
-        this.targetOffset = offset
-        const result = this.findNodeAndOffset(el);
-
-        if (result !== null) {
-          const [node, offsetIndex] = result;
-          requestAnimationFrame(() => {
-            if (node.nodeType === 3) {
-              // Text node
-              selection.setBaseAndExtent(node, offsetIndex, node, offsetIndex);
-            } else if (node.nodeType === 1 && node.childNodes.length > 0) {
-              // Element node with child nodes (e.g., <span>)
-              selection.setBaseAndExtent(node.childNodes[0], offsetIndex, node.childNodes[0], offsetIndex);
-            }
-          });
-        }
-      } else {
-        requestAnimationFrame(() => {
-        const selection = window.getSelection();
-        selection.removeAllRanges();
-        selection.setBaseAndExtent(el.firstChild, offset, el.firstChild, offset)
-        })
-      }
-      }
-  }
-
-  /**
-   * Helper for Regex Highlighter, find right node and index for current cursor position
-   * @param element HTMLElement
-   * @returns node: node with cursor, number: offest of cursor in node
-   */
-  findNodeAndOffset(element: Node): [Node, number] | null {
-    if (element.nodeType === 3) {
-      // Text node
-      const textLength = (element.nodeValue || "").length;
-      if (this.targetOffset <= textLength) {
-        return [element, this.targetOffset];
-      } else {
-        this.targetOffset -= textLength;
-      }
-    } else if (element.nodeType === 1){
-      // Element node
-      for (let i = 0; i < element.childNodes.length; i++) {
-        const child = element.childNodes[i];
-        const result = this.findNodeAndOffset(child);
-        if (result !== null) {
-          return result;
-        }
-      }
-    }
-    return null;
-  }
-
-  /**
-     * Helper for Regex Highlighter, extract current cursor position
-     * @param element HTMLElement
-     * @returns num, offset of cursor position
-     */
-  getCaretCharacterOffsetWithin(element) {
-    var caretOffset = 0;
-    var doc = element.ownerDocument || element.document;
-    var win = doc.defaultView || doc.parentWindow;
-    var sel;
-    if (typeof win.getSelection != "undefined") {
-        sel = win.getSelection();
-        if (sel.rangeCount > 0) {
-            var range = win.getSelection().getRangeAt(0);
-            var preCaretRange = range.cloneRange();
-            preCaretRange.selectNodeContents(element);
-            preCaretRange.setEnd(range.endContainer, range.endOffset);
-            caretOffset = preCaretRange.toString().length;
-        }
-    } else if ( (sel = doc.selection) && sel.type != "Control") {
-        var textRange = sel.createRange();
-        var preCaretTextRange = doc.body.createTextRange();
-        preCaretTextRange.moveToElementText(element);
-        preCaretTextRange.setEndPoint("EndToEnd", textRange);
-        caretOffset = preCaretTextRange.text.length;
-    }
-    return caretOffset;
   }
 
   /**
